+++ newrelic/core/function_node.py
@@ -1,16 +1,18 @@
 from collections import namedtuple
 
+import newrelic.core.attribute as attribute
 import newrelic.core.trace_node
 
 from newrelic.core.node_mixin import GenericNodeMixin
 from newrelic.core.metric import TimeMetric
+from newrelic.core.attribute_filter import DST_TRANSACTION_SEGMENTS
 
 from newrelic.packages import six
 
 _FunctionNode = namedtuple('_FunctionNode',
         ['group', 'name', 'children', 'start_time', 'end_time',
         'duration', 'exclusive', 'label', 'params', 'rollup',
-        'is_async', 'guid'])
+        'is_async', 'guid', 'agent_attributes'])
 
 
 class FunctionNode(_FunctionNode, GenericNodeMixin):
@@ -91,7 +93,17 @@ class FunctionNode(_FunctionNode, GenericNodeMixin):
                 break
             children.append(child.trace_node(stats, root, connections))
 
-        params = self.params or {}
+        # Agent attributes
+        params = attribute.resolve_agent_attributes(
+                self.agent_attributes,
+                root.settings.attribute_filter,
+                DST_TRANSACTION_SEGMENTS)
+
+        # User attributes override agent attributes
+        if self.params:
+            params.update(self.params)
+
+        # Intrinsic attributes override everything
         params['exclusive_duration_millis'] = 1000.0 * self.exclusive
 
         return newrelic.core.trace_node.TraceNode(start_time=start_time,