+++ newrelic/core/tests/test_datastore_node.py
@@ -34,6 +34,7 @@ _ds_node = newrelic.core.datastore_node.DatastoreNode(
         database_name='bar',
         is_async=False,
         guid=None,
+        agent_attributes={},
 )